Refactor: norm 맞춰 파일 정리
