+++ srcs/render.c
@@ -6,7 +6,7 @@
 /*   By: jihoh <jihoh@student.42seoul.kr>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/07/26 16:36:33 by jihoh             #+#    #+#             */
-/*   Updated: 2022/07/30 04:24:41 by jihoh            ###   ########.fr       */
+/*   Updated: 2022/07/30 04:48:59 by jihoh            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,186 +23,7 @@ t_ray	create_ray(t_cam *cam, double x, double y)
 	return (ray);
 }
 
-int	check_rgb(int n)
-{
-	if (n > 0xFF)
-		return (0xFF);
-	else if (n < 0)
-		return (0);
-	else
-		return (n);
-}
-
-int	cscale(int color, double d)
-{
-	int		rgb[3];
-
-	rgb[0] = check_rgb(d * (color >> 0x10));
-	rgb[1] = check_rgb(d * ((color >> 0x08) & 0xFF));
-	rgb[2] = check_rgb(d * (color & 0xFF));
-	return ((rgb[0] << 0x10) | (rgb[1] << 0x08) | rgb[2]);
-}
-
-int	cprod(int c1, int c2)
-{
-	int	rgb[3];
-
-	rgb[0] = (((float)(c1 >> 0x10) / 0xFF)
-			* ((float)(c2 >> 0x10) / 0xFF)) * 0xFF;
-	rgb[1] = (((float)((c1 >> 0x08) & 0xFF) / 0xFF)
-			* ((float)((c2 >> 0x08) & 0xFF) / 0xFF)) * 0xFF;
-	rgb[2] = (((float)(c1 & 0xFF) / 0xFF)
-			* ((float)(c2 & 0xFF) / 0xFF)) * 0xFF;
-	return ((rgb[0] << 0x10) | (rgb[1] << 0x08) | rgb[2]);
-}
-
-int	cadd(int c1, int c2)
-{
-	int		r;
-	int		g;
-	int		b;
-
-	r = check_rgb((c1 >> 0x10) + (c2 >> 0x10));
-	g = check_rgb((c1 >> 0x08 & 0xFF) + (c2 >> 0x08 & 0xFF));
-	b = check_rgb((c1 & 0xFF) + (c2 & 0xFF));
-	return ((r << 0x10) | (g << 0x08) | b);
-}
-
-int	ccomp(t_light *light, t_hit hit)
-{
-	t_vec3		light_normal;
-	float		gain;
-	float		r2;
-	float		light_bright;
-
-	light_normal = vsub(light->o, hit.point);
-	r2 = length_squared(light_normal);
-	gain = dot(normalize(light_normal), hit.nv);
-	if (gain <= 0)
-		light_bright = 0;
-	else
-		light_bright = (light->br * gain * ALBEDO) / (4.0 * M_PI * r2);
-	return (cprod(cadd(0, cscale(hit.clr, light_bright)), light->clr));
-}
-
-void	get_sphere_root(double root[2], t_ray *ray, t_sphere sp)
-{
-	double	discriminant;
-	t_p3	oc;
-	double	k[3];
-
-	oc = vsub(ray->o, sp.c);
-	k[0] = dot(ray->dir, ray->dir);
-	k[1] = 2 * dot(ray->dir, oc);
-	k[2] = dot(oc, oc) - sp.r * sp.r;
-	discriminant = k[1] * k[1] - (4 * k[0] * k[2]);
-	if (discriminant < 0)
-	{
-		root[0] = INFINITY;
-		root[1] = INFINITY;
-	}
-	else
-	{
-		root[0] = (-k[1] - sqrt(discriminant)) / (2 * k[0]);
-		root[1] = (-k[1] + sqrt(discriminant)) / (2 * k[0]);
-	}
-}
-
-t_p3	get_hit_point(t_ray ray)
-{
-	return (vadd(ray.o, vscale(ray.dir, ray.hit.time)));
-}
-
-int	hit_plane(t_figures *elem, t_ray *ray)
-{
-	double	time;
-	double	den;
-
-	den = dot(normalize(ray->dir), elem->nv);
-	if (!den)
-		return (0);
-	time = (dot(vsub(elem->fig.pl.p, ray->o), elem->nv) / den);
-	if (ray->hit.time > time && time > EPSILON)
-	{
-		ray->hit.time = time;
-		ray->hit.point = get_hit_point(*ray);
-		if (dot(ray->dir, elem->nv) > 0)
-			elem->nv = vscale(elem->nv, -1);
-		ray->hit.nv = elem->nv;
-		ray->hit.clr = elem->clr;
-		return (1);
-	}
-	return (0);
-}
-
-int	hit_sphere(t_figures *elem, t_ray *ray)
-{
-	double		time[2];
-	t_sphere	sp;
-
-	sp = elem->fig.sp;
-	get_sphere_root(time, ray, sp);
-	if (ray->hit.time > time[0] && time[0] > 0)
-	{
-		ray->hit.time = time[0];
-		ray->hit.point = get_hit_point(*ray);
-		ray->hit.nv = normalize(vsub(ray->hit.point, sp.c));
-		ray->hit.clr = elem->clr;
-		return (1);
-	}
-	return (0);
-}
-
-int	intersect(t_minirt *rt, t_ray *ray)
-{
-	int			ret;
-	t_figures	*elem;
-
-	ray->hit.time = INFINITY;
-	ret = 0;
-	elem = rt->scene.figures;
-	while (elem)
-	{
-		if (elem->type == SP)
-			ret |= hit_sphere(elem, ray);
-		else if (elem->type == PL)
-			ret |= hit_plane(elem, ray);
-		elem = elem->next;
-	}
-	return (ret);
-}
-
-int	in_shadow(t_minirt *rt, t_hit hit, t_light *light)
-{
-	t_ray	shadow;
-
-	shadow.o = hit.point;
-	shadow.dir = normalize(vsub(light->o, hit.point));
-	return (intersect(rt, &shadow));
-}
-
-int	raytrace(t_minirt *rt, t_ray *ray)
-{
-	int		al_color;
-	int		color;
-	int		vis;
-	t_light	*light;
-
-	if (!intersect(rt, ray))
-		return (0);
-	al_color = cscale(rt->scene.al_clr, rt->scene.al_br);
-	color = cprod(ray->hit.clr, al_color);
-	light = rt->scene.light;
-	while (light)
-	{
-		vis = !in_shadow(rt, ray->hit, light);
-		color = cadd(color, vis * ccomp(light, ray->hit));
-		light = light->next;
-	}
-	return (color);
-}
-
-void		mlx_put_pixel2img(t_img *img, int x, int y, int colour)
+void	set_pixel_color(t_img *img, int x, int y, int colour)
 {
 	char	*dst;
 
@@ -232,7 +53,7 @@ void	render_scene(t_minirt *rt, t_cam *cam)
 		{
 			ray = create_ray(cam, x / rt->scene.xres, y / rt->scene.yres);
 			color = raytrace(rt, &ray);
-			mlx_put_pixel2img(&cam->img, x, (rt->scene.yres - 1) - y, color);
+			set_pixel_color(&cam->img, x, (rt->scene.yres - 1) - y, color);
 		}
 	}
 	if (!rt->save)
